Debugged speciality from specialty.
